+++ src/pages/api/postDelete.ts
@@ -17,6 +17,9 @@ if (!admin.apps.length) {
 }
 
 const firestore = admin.firestore();
+const bucket = admin
+  .storage()
+  .bucket(process.env.NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET_ID as string);
 
 export default async function handler(
   req: NextApiRequest,
@@ -25,8 +28,28 @@ export default async function handler(
   if (req.method === "DELETE") {
     const { postDocPath } = req.body;
 
+    if (!postDocPath) {
+      res.status(405).json({ error: "Missing Prop" });
+      console.error("Missing Prop");
+      return;
+    }
+
     try {
       await firestore.doc(postDocPath).delete();
+
+      // Delete NFT Metadata if this post has. THIS PROCESS SHOULD BE AUTOMATED
+      const ps = await firestore.doc(postDocPath).get();
+      const data = ps.data();
+
+      if (data && data.nftUrl) {
+        const metadataPath = `users/${postDocPath.split("/")[1]}/nftMetadatas/${
+          postDocPath.split("/")[3]
+        }`;
+
+        const oldNftMetadataFile = bucket.file(metadataPath);
+        await oldNftMetadataFile.delete();
+      }
+
       res.status(200).json({});
     } catch (error) {
       console.error("Error while deleting doc", error);
